Toolchain: final conclusion
Although this toolchain actually seems to compile the resulting binaries
it produces are not compatible due to having a newer Glibc than what the
Boxee is running natively.
The only way the current toolchian seems to work is by using an
identical Glibc version (2.7). Maybe we can use this in the future if
the Boxee gets some attention by someone in the know to get a newer
kernel installed.