settings.xml: sync playback to display as default
This seems to stop the frames dropping.
When subtitles are enabled some frames are still dropped at the start (along with some gfx glitching), but the good news is everything settles down after a minute or two and no further frames get dropped and no further gfx glitches occur from then on.

I'm no coder, but what I'm think is happening is when subtitles are enabled this causes extra resources to be taken up. What with the video audio sync taking up most of the resources at the start of the video the subtitles kicking in (especially if there are a lot at the beginning), causes the gfx glitches to occur.

With these changes it goes some way to alleviate the subtitle glitching for Boxee Box users. This is all I can pull out of the bag for the moment.